Added MovieService and test

+++ Moviemo.IntegrationTests/Extensions/TestUtils.cs
@@ -7,6 +7,7 @@ namespace Moviemo.IntegrationTests.Extensions
     public static class TestUtils
     {
         public const int InvalidGenreId = 100;
+        public const int InvalidMovieId = 100;
 
         public static void Init (MoviemoContext context)
         {
@@ -16,6 +17,7 @@ namespace Moviemo.IntegrationTests.Extensions
             }
 
             context.Genres.AddRange(GetGenres());
+            context.Movies.AddRange(GetMovies());
             context.SaveChanges();
         }
 
@@ -26,8 +28,11 @@ namespace Moviemo.IntegrationTests.Extensions
                 throw new System.ArgumentNullException(nameof(context));
             }
 
-            if (context.Genres != null)
+            if (context.Genres != null && context.Movies != null)
+            {
                 context.Genres.RemoveRange(context.Genres);
+                context.Movies.RemoveRange(context.Movies);
+            }
 
             Init(context);
         }
@@ -40,6 +45,17 @@ namespace Moviemo.IntegrationTests.Extensions
                 new Genre() { Id = 2, Name = "Drama"},
                 new Genre() { Id = 3, Name = "Comedy"},
                 new Genre() { Id = 4, Name = "Romance"}
+            };
+        }
+
+        public static List<Movie> GetMovies ()
+        {
+            return new List<Movie>()
+            {
+                new Movie() { Id = 1, GenreId = 1, Title = "Avengers"},
+                new Movie() { Id = 2, GenreId = 2, Title = "The Rainmaker"},
+                new Movie() { Id = 3, GenreId = 3, Title = "Jumanji"},
+                new Movie() { Id = 4, GenreId = 4, Title = "Me Before You"}
             };
         }
